create: Player can walk, run and dash
Going to have to improve dashing as it may be too broken
